:recycle: refactor(ipsr readonly) LFUB-348: Fields and buttons read only validation added

+++ onecgiar-pr-client/src/app/shared/directives/ymz-list-structure-item/ymz-list-structure-item.directive.ts
@@ -1,11 +1,12 @@
 import { Directive, Output, EventEmitter, ElementRef, HostListener } from '@angular/core';
+import { RolesService } from '../../services/global/roles.service';
 
 @Directive({
   selector: '[appYmzListStructureItem]'
 })
 export class YmzListStructureItemDirective {
   @Output() deleteEvent = new EventEmitter();
-  constructor(private el: ElementRef) {
+  constructor(private el: ElementRef, private rolesSE: RolesService) {
     el.nativeElement.classList.add('ymz-list-structure-item');
   }
   @HostListener('click', ['$event']) onClick(event) {
@@ -18,14 +19,17 @@ export class YmzListStructureItemDirective {
     //Called after the constructor, initializing input properties, and the first call to ngOnChanges.
     //Add 'implements OnInit' to the class.
     const div = document.createElement('div');
-    div.classList.add('clickEvent', 'deleteItem');
-    const icon = document.createElement('i');
-    icon.classList.add('clickEvent', 'material-icons-round', 'ymz-lsi-delete-icon');
-    icon.innerText = 'delete';
-    div.appendChild(icon);
+
+    if (!this.rolesSE.readOnly) {
+      div.classList.add('clickEvent', 'deleteItem');
+      const icon = document.createElement('i');
+      icon.classList.add('clickEvent', 'material-icons-round', 'ymz-lsi-delete-icon');
+      icon.innerText = 'delete';
+      div.appendChild(icon);
+    }
     const line = document.createElement('div');
     line.classList.add('line');
     this.el.nativeElement.appendChild(line);
-    this.el.nativeElement.appendChild(div);
+    if (!this.rolesSE.readOnly) this.el.nativeElement.appendChild(div);
   }
 }
